+++ src/app/styles/open-app-button-style/open-app-button-style.component.spec.ts
@@ -1,0 +1,24 @@
+import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
+import { IonicModule } from '@ionic/angular';
+
+import { OpenAppButtonStyleComponent } from './open-app-button-style.component';
+
+describe('OpenAppButtonStyleComponent', () => {
+  let component: OpenAppButtonStyleComponent;
+  let fixture: ComponentFixture<OpenAppButtonStyleComponent>;
+
+  beforeEach(waitForAsync(() => {
+    TestBed.configureTestingModule({
+      declarations: [ OpenAppButtonStyleComponent ],
+      imports: [IonicModule.forRoot()]
+    }).compileComponents();
+
+    fixture = TestBed.createComponent(OpenAppButtonStyleComponent);
+    component = fixture.componentInstance;
+    fixture.detectChanges();
+  }));
+
+  it('should create', () => {
+    expect(component).toBeTruthy();
+  });
+});