+++ CHANGELOG.md
@@ -2,6 +2,7 @@
  - add [capacitor-native-settings](https://github.com/RaphaelWoude/capacitor-native-settings) v5.0.1
  - add style `shortcutButton` - open native window
  - add [capacitor/app-launcher](https://capacitorjs.com/docs/apis/app-launcher)
+ - add style `openAppButton` - open app based on the configuration, if the app is not installed it opens the installation link if there is one
 
 # 3.1.10
  - load `toggle switch` when it is enabled for a input from type `checkbox`